feat: unregister 에 user 유효성검사 추가

+++ server/src/modules/user/user.service.ts
@@ -1,11 +1,18 @@
 import { Injectable } from '@nestjs/common'
 import { UserRepository } from './user.repository'
+import { ExceptionError } from '@/shared/lib/errors'
 
 @Injectable()
 export class UserService {
   constructor(private readonly userRepository: UserRepository) {}
 
   async unregister(userId: number) {
+    const user = await this.userRepository.existsByUserId(userId)
+
+    if (!user) {
+      throw new ExceptionError('NotFound', 'User not found')
+    }
+
     await this.userRepository.deleteUser(userId)
   }
 }
